DATAJPA-1403 - Fixed failing Tests for Hibernate 5.2+.
The PersistenceProvider of Hibernate changed its name. Therefore we need to provide multiple persistence contexts for Hibernate 5 vs Hibernate 4.

Also, some tests were ignored for Hibernate 4 which now run with Hibernate 5 on the classpath. This happens even when the actual JPA Provider used is EclipseLink or OpenJPA.
Both don't have proper support for Tuples which is what is needed for those tests. Therefore the affected tests got added to the ever longer lists of ignored tests.

We do not execute OpenJpa Tests for spring-51-next as Spring 5 and above require a JPA version which is not supported by OpenJPA.

Original pull request: #289.